+++ src/main/java/de/schaack/ml/basics/functions/evaluations/classification/binary/BinaryEvaluationEnum.java
@@ -1,0 +1,5 @@
+package de.schaack.ml.basics.functions.evaluations.classification.binary;
+
+public enum BinaryEvaluationEnum {
+    TRUE_POSITIVE, TRUE_NEGATIVE, FALSE_POSITIVE, FALSE_NEGATIVE
+}